ref: Split welcome trial into WelcomeMessage and LargeWindow
