+++ src/JsPsych/trials/examples/index.js
@@ -2,5 +2,7 @@ export * from './AdjustVolume';
 export * from './AgeCheck';
 export * from './Countdown';
 export * from './EndExperiment';
+export * from './LargeWindow';
 export * from './Preamble';
 export * from './Slider';
+export * from './WelcomeMessage';
