--- src/trials/welcome.js
@@ -1,35 +1,0 @@
-import htmlKeyboardResponse from '@jspsych/plugin-html-keyboard-response';
-import { lang } from '../config/main';
-import { photodiodeGhostBox } from '../lib/markup/photodiode';
-import { baseStimulus } from '../lib/markup/stimuli';
-
-// TODO 162: Pull into its own file
-function pleaseBiggen() {
-  const stimulus =
-    baseStimulus(`<h1>${lang.welcome.large_window}</h1>`, true) + photodiodeGhostBox();
-
-  return {
-    type: htmlKeyboardResponse,
-    stimulus,
-    prompt: lang.prompt.continue.press,
-    response_ends_trial: true,
-  };
-}
-
-function welcomeMessage() {
-  const stimulus = baseStimulus(`<h1>${lang.welcome.message}</h1>`, true) + photodiodeGhostBox();
-
-  return {
-    type: htmlKeyboardResponse,
-    stimulus,
-    prompt: lang.prompt.continue.press,
-    response_ends_trial: true,
-  };
-}
-
-const welcome = {
-  type: htmlKeyboardResponse,
-  timeline: [pleaseBiggen(), welcomeMessage()],
-};
-
-export default welcome;
